doc_analyzer/requirements.txt
mistralai==0.0.7  # Cliente API Mistral
python-dotenv==1.0.0  # Para variáveis de ambiente
PyMuPDF==1.23.8  # Para processamento de PDFs
pytesseract==0.3.10  # Para OCR
Pillow==10.1.0  # Para processamento de imagens
customtkinter==5.2.1  # Para interface gráfica moderna
PyYAML==6.0.1  # Para arquivos de configuração
anthropic==0.7.7
openai==1.3.7
requests==2.31.0
python-docx==1.0.1
tqdm==4.66.1
colorama==0.4.6
